// api/src/user/interfaces/user.interface.ts
export interface UserInterface {
  readonly uuid: number
  readonly fName: string
  readonly lName: string
  readonly nickName: string
  readonly email: string
  readonly password: string
  readonly facebook: string
  readonly twitter: string
  readonly line: string
  readonly youtube: string
  readonly website: string
}